requirements.txt

#
# Requirements for wger for production
#
Django>=1.9,<1.10
django-recaptcha==1.3.1
reportlab
django_mobile
django-formtools>=1.0,<1.1
bleach
python-mimeparse
pillow
easy-thumbnails
django_compressor
icalendar
django-bootstrap-breadcrumbs
sphinx
django-sortedm2m
django-bower
invoke
requests
psycopg2-binary
whitenoise
gunicorn
dj-database-url==0.4.2
dj-static==0.0.6
social-auth-app-django

# REST API
djangorestframework>=3.2,<3.3
django-filter
django-tastypie
django-cors-headers

# Python3 compatibility
six
autopep8
flake8
